JvmEnvironment{Params,Result} => JvmTestEnvironment{Params,Result}
